requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
twilio==9.1.0
sqlalchemy==2.0.30
# psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
